requirements.txt
# PySpark and dependencies
pyspark>=3.4.0
py4j>=0.10.9.7

# Configuration and data handling
PyYAML>=6.0
pandas>=1.5.0

# Testing
pytest>=7.0.0
pytest-spark>=0.6.0

# Development tools
black>=22.0.0
flake8>=5.0.0

# Optional: For Avro support
# avro-python3>=1.10.0

# Optional: For enhanced logging
colorlog>=6.0.0